[refactor] Extract whisper_api package
