+++ speech_recognition/recognizers/whisper_api/base.py
