+++ tests/recognizers/test_groq.py
@@ -4,7 +4,7 @@ import httpx
 import respx
 
 from speech_recognition import AudioData, Recognizer
-from speech_recognition.recognizers import groq
+from speech_recognition.recognizers.whisper_api import groq
 
 
 @respx.mock(assert_all_called=True, assert_all_mocked=True)
